Honour type promotion on variable argument handling
* linux/dev/lib/vsprintf.c (linux_vsprintf): Use `int' when retrieving
a `short' through `va_arg'.

+++ linux/dev/lib/vsprintf.c
@@ -328,9 +328,9 @@ linux_vsprintf (char *buf, const char *fmt, va_list args)
 	num = va_arg (args, unsigned long);
       else if (qualifier == 'h')
 	if (flags & SIGN)
-	  num = va_arg (args, short);
+	  num = (short) va_arg (args, int);
 	else
-	  num = va_arg (args, unsigned short);
+	  num = (unsigned short) va_arg (args, unsigned int);
       else if (flags & SIGN)
 	num = va_arg (args, int);
       else
